Restructured transit project
