+++ example_apps/transit/Cargo.toml
@@ -15,3 +15,4 @@ percent-encoding = { workspace = true }
 reqwest = { workspace = true, features = ["rustls-tls", "gzip"]}
 bytes = { workspace = true }
 tracing = { workspace = true }
+transit-model = { path = "transit-model" }
