+++ example_apps/transit/transit-model/src/model/agency.rs
@@ -31,19 +31,6 @@ pub struct Agency {
     pub country: String,
 }
 
-const AGENCIES_CSV: &[u8] = include_bytes!("agencies.csv");
-
-pub fn agencies() -> Vec<Agency> {
-    let mut reader = csv::Reader::from_reader(AGENCIES_CSV);
-    let agencies_result = reader.deserialize::<Agency>().collect::<Result<_, _>>();
-
-    let mut agencies: Vec<Agency> = agencies_result.expect("CSV data was invalid.");
-    for (i, agency) in agencies.iter_mut().enumerate() {
-        agency.index = i;
-    }
-    agencies
-}
-
 fn enc<'a>(s: &'a str) -> PercentEncode<'a> {
     percent_encoding::utf8_percent_encode(s, NON_ALPHANUMERIC)
 }
@@ -100,16 +87,5 @@ impl Agency {
             predictable,
             route_title: route.title.clone(),
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-
-    #[test]
-    fn load_agencies() {
-        let agencies = super::agencies();
-        println!("{:?}", agencies);
-        assert_eq!(agencies.len(), 46);
     }
 }
