+++ example_apps/transit/transit-model/src/model/bounding_box.rs
