+++ example_apps/transit/transit-model/src/model/counts.rs
