+++ example_apps/transit/transit-model/src/model/test-data/vehicles.xml
