+++ example_apps/transit/transit-model/src/model/vehicle.rs
